requirements.txt
amqp==2.5.2
atomicwrites==1.3.0
attrs==19.3.0
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coverage==5.0.3
dj-database-url==0.5.0
Django==2.2.10
django-celery-beat==1.6.0
django-celerybeat-status==0.0.10
django-grappelli==2.14.1
django-heroku==0.3.1
django-timezone-field==3.1
djangorestframework==3.11.0
idna==2.9
importlib-metadata==1.5.0
kombu==4.6.7
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.8.1
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
whitenoise==5.0.1
zipp==3.0.0
